move parSequence safety to JVM-only
